next-app/scripts/requirements.txt
playwright
beautifulsoup4 
pymongo
python-dotenv